refactor: icrc transactions utils
