+++ frontend/src/lib/components/accounts/IcrcTransactionCard.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
   import type { Account } from "$lib/types/account";
-  import { mapSnsTransaction } from "$lib/utils/sns-transactions.utils";
-  import type { Transaction } from "$lib/utils/transactions.utils";
+  import { mapIcrcTransaction } from "$lib/utils/icrc-transactions.utils";
   import type { Principal } from "@dfinity/principal";
   import type { IcrcTransactionWithId } from "@dfinity/ledger";
   import TransactionCard from "./TransactionCard.svelte";
+  import type { Transaction } from "$lib/types/transaction";
 
   export let transactionWithId: IcrcTransactionWithId;
   export let account: Account;
@@ -12,7 +12,7 @@
   export let governanceCanisterId: Principal | undefined = undefined;
 
   let transactionData: Transaction | undefined;
-  $: transactionData = mapSnsTransaction({
+  $: transactionData = mapIcrcTransaction({
     transaction: transactionWithId,
     account,
     toSelfTransaction,
