+++ frontend/src/lib/types/transaction.ts
@@ -1,3 +1,5 @@
+import type { IcrcTransactionWithId } from "@dfinity/ledger";
+import type { TokenAmount } from "@dfinity/nns";
 import type { Account } from "./account";
 
 export type NewTransaction = {
@@ -9,3 +11,42 @@ export type NewTransaction = {
 export type ValidateAmountFn = (
   amount: number | undefined
 ) => string | undefined;
+
+export interface IcrcTransactionData {
+  toSelfTransaction: boolean;
+  transaction: IcrcTransactionWithId;
+}
+
+export interface IcrcTransactionInfo {
+  to?: string;
+  from?: string;
+  memo?: Uint8Array;
+  created_at_time?: bigint;
+  amount: bigint;
+  fee?: bigint;
+}
+
+// Value should match the key in i18n "transaction_names"
+export enum AccountTransactionType {
+  Burn = "burn",
+  Mint = "mint",
+  Send = "send",
+  StakeNeuron = "stakeNeuron",
+  StakeNeuronNotification = "stakeNeuronNotification",
+  TopUpNeuron = "topUpNeuron",
+  CreateCanister = "createCanister",
+  TopUpCanister = "topUpCanister",
+  ParticipateSwap = "participateSwap",
+}
+
+export interface Transaction {
+  type: AccountTransactionType;
+  isReceive: boolean;
+  isSend: boolean;
+  // Account string representation
+  from: string | undefined;
+  // Account string representation
+  to: string | undefined;
+  displayAmount: TokenAmount;
+  date: Date;
+}
