+++ frontend/src/tests/mocks/sns-transactions.mock.ts
@@ -30,34 +30,6 @@ export const mockSnsTransactionWithId: IcrcTransactionWithId = {
   transaction: mockSnsTransaction,
 };
 
-interface SnsCandidAccount {
-  owner: Principal;
-  subaccount: [] | [Uint8Array];
-}
-
-export const createSnstransactionWithId = (
-  to: SnsCandidAccount,
-  from: SnsCandidAccount
-): IcrcTransactionWithId => ({
-  id: BigInt(123),
-  transaction: {
-    kind: "transfer",
-    timestamp: BigInt(12354),
-    burn: [],
-    mint: [],
-    transfer: [
-      {
-        to,
-        from,
-        memo: [],
-        created_at_time: [BigInt(123)],
-        amount: BigInt(33),
-        fee: [BigInt(1)],
-      },
-    ],
-  },
-});
-
 export const mockSnsTransactionsStoreSubscribe =
   (store: SnsTransactionsStoreData) =>
   (run: Subscriber<SnsTransactionsStoreData>): (() => void) => {
