+++ frontend/src/tests/mocks/transaction.mock.ts
@@ -1,8 +1,6 @@
 import type { Transaction as NnsTransaction } from "$lib/canisters/nns-dapp/nns-dapp.types";
-import {
-  AccountTransactionType,
-  type Transaction,
-} from "$lib/utils/transactions.utils";
+import type { Transaction } from "$lib/types/transaction";
+import { AccountTransactionType } from "$lib/types/transaction";
 import { ICPToken, TokenAmount } from "@dfinity/nns";
 import { mockMainAccount, mockSubAccount } from "./accounts.store.mock";
